fix failing specs
this is fairly large. After we made the critical changes, a number of tests broke. We fixed those. As we addressed each test we also discovered some opportunities to add additional tests. We also abstracted some new scopes, for easier access to things that were previously not abstracted enough (they knew too much about the associations).